Added constants, reworked intents from tabs, edited date fields in model

+++ app/src/main/java/com/example/a/model/repository/LinkRepository.java
@@ -76,7 +76,7 @@ public class LinkRepository {
         );
     }
 
-    public void deleteByIdRxJava(Long id) {
+    public void deleteByIdRxJava(int id) {
         mDisposable.add(deleteLinkById(id)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
